requirement.txt
pytorch==2.1.1
transformers==4.35.1
pytorch_lightning==2.1.3
datasets
ftfy
regex
tqdm
sentencepiece